res fixed some casts in hashmap.cpp causing grief on 64bit systems.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18503 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/hashmap.cpp
@@ -32,8 +32,8 @@ csGlobalHashIterator::csGlobalHashIterator (csHashMap *hm)
   cbucket = bucket = 0;
   bucket_len = 0;
   element_index = 0;
-  bucket_index = (unsigned int)-1;
-  nbuckets = (unsigned int)hash->Buckets.Length();
+  bucket_index = (size_t)-1;
+  nbuckets = hash->Buckets.Length();
   GotoNextElement ();
 }
